Use correct type for looping data in CRC calculation

+++ src/EE895.cpp
@@ -6,7 +6,7 @@ EE895::EE895(){
 uint16_t EE895::calculateCRC(const unsigned char* data, size_t length){
   uint16_t crc = 0xFFFF;
 
-  for (int pos = 0; pos < length; pos++) {
+  for (size_t pos = 0; pos < length; pos++) {
     crc ^= (uint16_t)data[pos];
     for (uint8_t bit = 8; bit > 0; bit--) {
       if ((crc & 0x0001) != 0) {
